long policy
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5630 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/panic.c
@@ -77,7 +77,7 @@ void panicf( const char *fmt, ...)
     while (1)
     {
 #ifndef SIMULATOR
-        volatile int i;
+        volatile long i;
         led (state);
         state = state?false:true;
